move subcommands to each subdirs.
