+++ command/daily/command.go
@@ -1,10 +1,11 @@
-package command
+package daily
 
 import (
 	"fmt"
 	"strings"
 	"time"
 
+	base "github.com/hirakiuc/daily-workflow/command"
 	"github.com/hirakiuc/daily-workflow/service"
 	"github.com/pkg/errors"
 	cli "github.com/urfave/cli/v2"
@@ -14,15 +15,15 @@ const JSTDiffSeconds = 9 * 60 * 60
 
 var ErrUnsupportedCase = errors.New("unsupported case: can't select multiple items")
 
-type DailyCommand struct {
+type Command struct {
 	Date time.Time
 
-	*Base
+	*base.Base
 }
 
-func NewDailyCommand(iostream *IoStream) *cli.Command {
-	srv := DailyCommand{
-		Base: NewBase(iostream),
+func NewCommand(iostream *base.IoStream) *cli.Command {
+	srv := Command{
+		Base: base.NewBase(iostream),
 	}
 
 	t := time.Now()
@@ -51,7 +52,7 @@ func NewDailyCommand(iostream *IoStream) *cli.Command {
 	}
 }
 
-func makeDailySubCommands(srv DailyCommand) []*cli.Command {
+func makeDailySubCommands(srv Command) []*cli.Command {
 	return []*cli.Command{
 		{
 			Name:   "edit",
@@ -90,10 +91,10 @@ func makeDailySubCommands(srv DailyCommand) []*cli.Command {
 	}
 }
 
-func (s *DailyCommand) parseArgs(c *cli.Context) error {
+func (s *Command) parseArgs(c *cli.Context) error {
 	err := s.LoadConfig("./config.toml")
 	if err != nil {
-		return err
+		return errors.Wrap(err, "Failed to load config")
 	}
 
 	d := c.String("date")
@@ -106,7 +107,7 @@ func (s *DailyCommand) parseArgs(c *cli.Context) error {
 	return nil
 }
 
-func (s *DailyCommand) EditAction(c *cli.Context) error {
+func (s *Command) EditAction(c *cli.Context) error {
 	err := s.parseArgs(c)
 	if err != nil {
 		return err
@@ -130,7 +131,7 @@ func (s *DailyCommand) EditAction(c *cli.Context) error {
 	return cmd.EditAndWait(fPath, "")
 }
 
-func (s *DailyCommand) findCandidates(_ *cli.Context, words []string) ([]string, error) {
+func (s *Command) findCandidates(_ *cli.Context, words []string) ([]string, error) {
 	fs := service.NewFsService(s.Conf.Common.Root)
 
 	if len(words) == 0 {
@@ -150,7 +151,7 @@ const (
 	OnlyOneItem             int = 1
 )
 
-func (s *DailyCommand) chooseAndEdit(_ *cli.Context, candidates []string) error {
+func (s *Command) chooseAndEdit(_ *cli.Context, candidates []string) error {
 	srv := service.NewCmdService(s.Conf)
 
 	results, err := srv.Choose(candidates)
@@ -183,7 +184,7 @@ func (s *DailyCommand) chooseAndEdit(_ *cli.Context, candidates []string) error 
 	return srv.EditAndWait(target, opts)
 }
 
-func (s *DailyCommand) ListAction(c *cli.Context) error {
+func (s *Command) ListAction(c *cli.Context) error {
 	err := s.parseArgs(c)
 	if err != nil {
 		return err
@@ -205,7 +206,7 @@ func (s *DailyCommand) ListAction(c *cli.Context) error {
 	return nil
 }
 
-func (s *DailyCommand) FindAction(c *cli.Context) error {
+func (s *Command) FindAction(c *cli.Context) error {
 	if err := s.parseArgs(c); err != nil {
 		return err
 	}
